+++ packages/app-defaults/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/app-defaults
 
+## 0.1.3
+
+### Patch Changes
+
+- 7927005152: Add `FetchApi` and related `fetchApiRef` which implement fetch, with an added Backstage token header when available.
+- 1e49c23eb7: Added an instance of PermissionApi to the apis included by default in createApp.
+- 68f8b10ccd: - Removed deprecation configuration option `theme` from `AppTheme` of the `AppThemeApi`
+  - Removed reference to `theme` in the `app-defaults` default `AppTheme`
+  - Removed logic in `AppThemeProvider` that creates `ThemeProvider` from `appTheme.theme`
+- Updated dependencies
+  - @backstage/core-app-api@0.3.0
+  - @backstage/core-plugin-api@0.4.0
+  - @backstage/plugin-permission-react@0.2.0
+  - @backstage/core-components@0.8.2
+
 ## 0.1.2
 
 ### Patch Changes