+++ plugins/cost-insights/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-cost-insights
 
+## 0.11.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-plugin-api@0.4.0
+  - @backstage/core-components@0.8.2
+
 ## 0.11.13
 
 ### Patch Changes