+++ plugins/search-backend-module-pg/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-search-backend-module-pg
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.10.0
+
 ## 0.2.2
 
 ### Patch Changes